+++ packages/react/src/index.ts
@@ -1,3 +1,10 @@
+export { ActiveProps } from './Active';
+export { BlockProps } from './Block';
+export { CuriBaseProps } from './CuriBase';
+export { CuriousProps, CuriousComponent } from './curious';
+export { CuriProps, CuriContext } from './interface';
+export { LinkProps, LinkState, ActiveLink } from './Link';
+
 import Active from './Active';
 import Block from './Block';
 import curious from './curious';